bug fix: the memo to share stack is now preserved if a terminal is parsing no
input. This was necessary for things like greedy that could parse no input.

As a consequence, terminal parsing nothing are now OK (except that they may still
be dangerous as they allow writing grammar that loops, like propably 'a*$*'.
This is probably not too bad, as this kind of things seems meaningless.